+++ solaris/utils/core.py
@@ -5,12 +5,13 @@ from shapely.geometry import Point
 from shapely.geometry.base import BaseGeometry
 import pandas as pd
 import geopandas as gpd
+import pyproj
 import rasterio
+from distutils.version import LooseVersion
 import skimage
 from fiona._err import CPLE_OpenFailedError
 from fiona.errors import DriverError
 from warnings import warn
-import pdb
 
 
 def _check_rasterio_im_load(im):
@@ -84,34 +85,19 @@ def _check_geom(geom):
         return Point(geom)
 
 
-def _check_crs(input_crs):
-    """Convert CRS to the integer format passed by ``solaris``."""
-    if isinstance(input_crs, dict):
-        # assume it's an {'init': 'epsgxxxx'} dict
-        out_crs = int(input_crs['init'].lower().strip('epsg:'))
-        out_crs = rasterio.crs.CRS.from_epsg(out_crs)
-    elif isinstance(input_crs, str):
-        #pdb.set_trace()
-        # handle PROJ4 strings, epsg strings, wkt strings
-        # but FIRST, see if it's just a number represented as a string
-        try:
-            input_crs = int(input_crs)
-            out_crs = rasterio.crs.CRS.from_epsg(input_crs)
-        except ValueError:
-            try:
-                out_crs = rasterio.crs.CRS.from_string(input_crs)
-            except rasterio.errors.CRSError as e:
-                raise ValueError(
-                    f"Solaris doesn't know how to parse {input_crs} as a "
-                    "crs. Try re-formatting. If this is properly formatted, "
-                    "open an issue in solaris's GitHub repository."
-                    ) from e
-    elif isinstance(input_crs, rasterio.crs.CRS):
+def _check_crs(input_crs, return_rasterio=False):
+    """Convert CRS to the ``pyproj.CRS`` object passed by ``solaris``."""
+    if not isinstance(input_crs, pyproj.CRS) and input_crs is not None:
+        out_crs = pyproj.CRS(input_crs)
+    else:
         out_crs = input_crs
-    elif isinstance(input_crs, int):
-        out_crs = rasterio.crs.CRS.from_epsg(input_crs)
-    elif input_crs is None:
-        out_crs = input_crs
+
+    if return_rasterio:
+        if LooseVersion(rasterio.__gdal_version__) >= LooseVersion("3.0.0"):
+            out_crs = rasterio.crs.CRS.from_wkt(out_crs.to_wkt())
+        else:
+            out_crs = rasterio.crs.CRS.from_wkt(out_crs.to_wkt("WKT1_GDAL"))
+
     return out_crs