+++ solaris/vector/mask.py
@@ -13,6 +13,8 @@ import rasterio
 from rasterio import features
 from affine import Affine
 from skimage.morphology import square, erosion, dilation
+import os
+from tqdm import tqdm
 
 def df_to_px_mask(df, channels=['footprint'], out_file=None, reference_im=None,
                   geom_col='geometry', do_transform=None, affine_obj=None,
@@ -889,6 +891,8 @@ def instance_mask(df, out_file=None, reference_im=None, geom_col='geometry',
         The value to use for nodata pixels in the mask. Defaults to 0 (the
         min value for ``uint8`` arrays). Used if reference_im nodata value is a float.
         Ignored if reference_im nodata value is an int or if reference_im is not used.
+        Take care when visualizing these masks, the nodata value may cause labels to not 
+        be visualized if nodata values are automatically masked by the software.
 
     Returns
     -------
@@ -904,7 +908,9 @@ def instance_mask(df, out_file=None, reference_im=None, geom_col='geometry',
             'If saving output to file, `reference_im` must be provided.')
     df = _check_df_load(df)
 
-    if len(df) == 0:
+    if len(df) == 0: # for saving an empty mask.
+        reference_im = _check_rasterio_im_load(reference_im)
+        shape = reference_im.shape
         return np.zeros(shape=shape, dtype='uint8')
 
     if do_transform is None:
@@ -948,9 +954,10 @@ def instance_mask(df, out_file=None, reference_im=None, geom_col='geometry',
     if reference_im:
         reference_im = _check_rasterio_im_load(reference_im)
     try:
-            bad_data_mask = (reference_im.read() == reference_im.nodata).any(axis=0) # take logical and along all dims so that all pixxels not -9999 across bands
+        bad_data_mask = (reference_im.read() == reference_im.nodata).any(axis=0) # take logical and along all dims so that all pixxels not -9999 across bands
     except AttributeError as ae:  # raise another, more verbose AttributeError
         raise AttributeError("A nodata value is not defined for the source image. Make sure the reference_im has a nodata value defined.") from ae
+    if len(bad_data_mask.shape) > 2:
         bad_data_mask = np.dstack([bad_data_mask]*output_arr.shape[2])
         output_arr = np.where(bad_data_mask, 0, output_arr) # mask is broadcasted to filter labels where there are non-nan image values
 
@@ -967,3 +974,60 @@ def instance_mask(df, out_file=None, reference_im=None, geom_col='geometry',
             dst.close()
 
     return output_arr
+
+def geojsons_to_masks_and_fill_nodata(rtiler, vtiler, label_tile_dir, fill_value=0):
+    """
+    Converts tiled vectors to raster labels and fills nodata values in raster and vector tiles.
+    
+    This function must be run after a raster tiler and vector tiler have already been initialized 
+    and the `.tile()` method for each has been called to generate raster and vector tiles. 
+    Geojson labels are first converted to rasterized masks, then the labels are set to 0 
+    where the reference image, the corresponding image tile, has nodata values. Then, nodata 
+    areas in the image tile are filled  in place with the fill_value. Only works for rasterizing 
+    all geometries as a single category with a burn value of 1. See test_tiler_fill_nodata in 
+    tests/test_tile/test_tile.py for an example.
+
+    Args
+    -------
+    rtiler : RasterTiler
+        The RasterTiler that has had it's `.tile()` method called.
+    vtiler : VectorTiler
+        The VectorTiler that has had it's `.tile()` method called.
+    label_tile_dir : str
+        The folder path to save rasterized labels. This is created if it doesn't already exist.
+    fill_value : str, optional
+        The value to use to fill nodata values in images. Defaults to 0.
+
+    Returns
+    -------
+    rasterized_label_paths : list
+        A list of the paths to the rasterized instance masks.
+    """
+    rasterized_label_paths = []
+    print("starting label mask generation")
+    if not os.path.exists(label_tile_dir):
+        os.mkdir(label_tile_dir)
+    for img_tile, geojson_tile in tqdm(zip(sorted(rtiler.tile_paths), sorted(vtiler.tile_paths))):
+        fid = os.path.basename(geojson_tile).split(".geojson")[0]
+        rasterized_label_path = os.path.join(label_tile_dir, fid + ".tif")
+        rasterized_label_paths.append(rasterized_label_path)
+        gdf = gpd.read_file(geojson_tile)
+        # gdf.crs = rtiler.raster_bounds_crs # add this because gdfs can't be saved with wkt crs
+        arr = instance_mask(gdf, out_file=rasterized_label_path, reference_im=img_tile, 
+                                        geom_col='geometry', do_transform=None,
+                                        out_type='int', burn_value=1, burn_field=None) # this saves the file, unless it is empty in which case we deal with it below.
+        if not arr.any(): # in case no instances in a tile we save it with "empty" at the front of the basename
+            with rasterio.open(img_tile) as reference_im:
+                meta = reference_im.meta.copy()
+                reference_im.close()
+            meta.update(count=1)
+            meta.update(dtype='uint8')
+            if isinstance(meta['nodata'], float):
+                meta.update(nodata=0)
+            rasterized_label_path = os.path.join(label_tile_dir, "empty_" + fid + ".tif")
+            with rasterio.open(rasterized_label_path, 'w', **meta) as dst:
+                dst.write(np.expand_dims(arr, axis=0))
+                dst.close()
+    rtiler.fill_all_nodata(nodata_fill=fill_value)
+    return rasterized_label_paths
+